+++ Sources/TuistGenerator/Generator/GenerationOptions.swift
@@ -1,8 +1,0 @@
-import Basic
-import Foundation
-
-/// Generation options.
-public class GenerationOptions {
-    /// Initializes the options with its attributes.
-    public init() {}
-}