+++ Sources/TuistGenerator/Graph/GraphLoaderCache.swift
@@ -11,15 +11,26 @@ protocol GraphLoaderCaching: AnyObject {
     func precompiledNode(_ path: AbsolutePath) -> PrecompiledNode?
     func add(targetNode: TargetNode)
     func targetNode(_ path: AbsolutePath, name: String) -> TargetNode?
+    func tuistConfig(_ path: AbsolutePath) -> TuistConfig?
+    func add(tuistConfig: TuistConfig, path: AbsolutePath)
 }
 
 /// Graph loader cache.
 class GraphLoaderCache: GraphLoaderCaching {
     // MARK: - GraphLoaderCaching
 
+    var tuistConfigs: [AbsolutePath: TuistConfig] = [:]
     var projects: [AbsolutePath: Project] = [:]
     var precompiledNodes: [AbsolutePath: PrecompiledNode] = [:]
     var targetNodes: [AbsolutePath: [String: TargetNode]] = [:]
+
+    func tuistConfig(_ path: AbsolutePath) -> TuistConfig? {
+        return tuistConfigs[path]
+    }
+
+    func add(tuistConfig: TuistConfig, path: AbsolutePath) {
+        tuistConfigs[path] = tuistConfig
+    }
 
     func project(_ path: AbsolutePath) -> Project? {
         return projects[path]