+++ Tests/TuistGeneratorTests/Generator/Mocks/MockWorkspaceGenerator.swift
@@ -5,14 +5,13 @@ import TuistCore
 
 final class MockWorkspaceGenerator: WorkspaceGenerating {
     var generateWorkspaces: [Workspace] = []
-    var generateStub: ((Workspace, AbsolutePath, Graphing, GenerationOptions, GenerationDirectory) throws -> AbsolutePath)?
+    var generateStub: ((Workspace, AbsolutePath, Graphing, TuistConfig) throws -> AbsolutePath)?
 
     func generate(workspace: Workspace,
                   path: AbsolutePath,
                   graph: Graphing,
-                  options: GenerationOptions,
-                  directory: GenerationDirectory) throws -> AbsolutePath {
+                  tuistConfig: TuistConfig) throws -> AbsolutePath {
         generateWorkspaces.append(workspace)
-        return (try generateStub?(workspace, path, graph, options, directory)) ?? AbsolutePath("/test")
+        return (try generateStub?(workspace, path, graph, tuistConfig)) ?? AbsolutePath("/test")
     }
 }