+++ Tests/TuistGeneratorTests/Models/TestData/TuistConfig+TestData.swift
@@ -1,0 +1,9 @@
+import Basic
+import Foundation
+@testable import TuistGenerator
+
+extension TuistConfig {
+    static func test(generationOptions: [GenerationOption] = []) -> TuistConfig {
+        return TuistConfig(generationOptions: generationOptions)
+    }
+}