+++ Tests/TuistKitTests/Commands/GenerateCommandTests.swift
@@ -99,7 +99,7 @@ final class GenerateCommandTests: XCTestCase {
         manifestLoader.manifestsAtStub = { _ in
             Set([.project])
         }
-        generator.generateProjectStub = { path, _, _ in
+        generator.generateProjectStub = { path, _ in
             generationPath = path
             return path.appending(component: "project.xcworkspace")
         }
@@ -118,7 +118,7 @@ final class GenerateCommandTests: XCTestCase {
         manifestLoader.manifestsAtStub = { _ in
             Set([.project])
         }
-        generator.generateProjectStub = { path, _, _ in
+        generator.generateProjectStub = { path, _ in
             generationPath = path
             return path.appending(component: "project.xcworkspace")
         }
@@ -137,7 +137,7 @@ final class GenerateCommandTests: XCTestCase {
         manifestLoader.manifestsAtStub = { _ in
             Set([.project])
         }
-        generator.generateProjectStub = { path, _, _ in
+        generator.generateProjectStub = { path, _ in
             generationPath = path
             return path.appending(component: "project.xcworkspace")
         }
@@ -170,7 +170,7 @@ final class GenerateCommandTests: XCTestCase {
         manifestLoader.manifestsAtStub = { _ in
             Set([.project])
         }
-        generator.generateProjectStub = { _, _, _ in
+        generator.generateProjectStub = { _, _ in
             throw error
         }