+++ Tests/TuistKitTests/Generator/GeneratorModelLoaderTests.swift
@@ -40,7 +40,7 @@ class GeneratorModelLoaderTest: XCTestCase {
     }
 
     override func tearDown() {
-        // Put teardown code here. This method is called after the invocation of each test method in the class.
+        fileHandler = nil
     }
 
     func test_loadProject() throws {
@@ -88,6 +88,38 @@ class GeneratorModelLoaderTest: XCTestCase {
         assert(target: model.targets[0], matches: targetA, at: path)
         assert(target: model.targets[1], matches: targetB, at: path)
         XCTAssertEqual(model.targets[2].name, "Project-Manifest")
+    }
+
+    func test_loadProject_withManifestTargetOptionDisabled() throws {
+        // Given
+        try fileHandler.createFiles([
+            "TuistConfig.swift",
+        ])
+        let projects = [
+            path: ProjectManifest.test(name: "Project",
+                                       targets: [
+                                           .test(name: "A"),
+                                           .test(name: "B"),
+                                       ]),
+        ]
+
+        let configs = [
+            path: TuistConfig.test(generationOptions: []),
+        ]
+
+        let manifestLoader = createManifestLoader(with: projects, configs: configs)
+        let subject = GeneratorModelLoader(fileHandler: fileHandler,
+                                           manifestLoader: manifestLoader,
+                                           manifestTargetGenerator: manifestTargetGenerator)
+
+        // When
+        let model = try subject.loadProject(at: path)
+
+        // Then
+        XCTAssertEqual(model.targets.map(\.name), [
+            "A",
+            "B",
+        ])
     }
 
     func test_loadProject_withAdditionalFiles() throws {
@@ -495,7 +527,8 @@ class GeneratorModelLoaderTest: XCTestCase {
 
     // MARK: - Helpers
 
-    func createManifestLoader(with projects: [AbsolutePath: ProjectDescription.Project]) -> GraphManifestLoading {
+    func createManifestLoader(with projects: [AbsolutePath: ProjectDescription.Project],
+                              configs: [AbsolutePath: ProjectDescription.TuistConfig] = [:]) -> GraphManifestLoading {
         let manifestLoader = MockGraphManifestLoader()
         manifestLoader.loadProjectStub = { path in
             guard let manifest = projects[path] else {
@@ -503,11 +536,22 @@ class GeneratorModelLoaderTest: XCTestCase {
             }
             return manifest
         }
-        manifestLoader.manifestsAtStub = { path in
-            guard let _ = projects[path] else {
-                return Set([])
+        manifestLoader.loadTuistConfigStub = { path in
+            guard let manifest = configs[path] else {
+                throw GraphManifestLoaderError.manifestNotFound(path)
             }
-            return Set([.project])
+            return manifest
+        }
+        manifestLoader.manifestsAtStub = { path in
+            var manifests = Set<Manifest>()
+            if projects[path] != nil {
+                manifests.insert(.project)
+            }
+
+            if configs[path] != nil {
+                manifests.insert(.tuistConfig)
+            }
+            return manifests
         }
         return manifestLoader
     }