+++ Tests/TuistKitTests/Generator/GraphManifestLoaderTests.swift
@@ -40,6 +40,25 @@ final class GraphManifestLoaderTests: XCTestCase {
                                       deprecator: deprecator)
     }
 
+    func test_loadTuistConfig() throws {
+        // Given
+        let content = """
+        import ProjectDescription
+        let config = TuistConfig(generationOptions: [.generateManifest])
+        """
+
+        let manifestPath = fileHandler.currentPath.appending(component: Manifest.tuistConfig.fileName)
+        try content.write(to: manifestPath.url,
+                          atomically: true,
+                          encoding: .utf8)
+
+        // When
+        let got = try subject.loadTuistConfig(at: fileHandler.currentPath)
+
+        // Then
+        XCTAssertTrue(got.generationOptions.contains(.generateManifest))
+    }
+
     func test_loadProject() throws {
         // Given
         let content = """
@@ -151,6 +170,7 @@ final class GraphManifestLoaderTests: XCTestCase {
         try fileHandler.touch(fileHandler.currentPath.appending(component: "Project.swift"))
         try fileHandler.touch(fileHandler.currentPath.appending(component: "Workspace.swift"))
         try fileHandler.touch(fileHandler.currentPath.appending(component: "Setup.swift"))
+        try fileHandler.touch(fileHandler.currentPath.appending(component: "TuistConfig.swift"))
 
         // When
         let got = subject.manifests(at: fileHandler.currentPath)
@@ -159,5 +179,6 @@ final class GraphManifestLoaderTests: XCTestCase {
         XCTAssertTrue(got.contains(.project))
         XCTAssertTrue(got.contains(.workspace))
         XCTAssertTrue(got.contains(.setup))
+        XCTAssertTrue(got.contains(.tuistConfig))
     }
 }