+++ Tests/TuistKitTests/Generator/TestData/ProjectDescription+TestData.swift
@@ -1,6 +1,12 @@
 import Foundation
 @testable import ProjectDescription
 
+extension TuistConfig {
+    static func test(generationOptions: [TuistConfig.GenerationOption] = []) -> TuistConfig {
+        return TuistConfig(generationOptions: generationOptions)
+    }
+}
+
 extension Workspace {
     static func test(name: String = "Workspace",
                      projects: [String] = [],